api/requirements.txt
fastapi
uvicorn
pydantic
alpaca-py
